requirements.txt
aioamqp==0.14.0
aiohttp==3.6.2
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
chardet==3.0.4
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.3.0
multidict==4.7.5
packaging==20.3
pamqp==2.3.0
pluggy==0.13.1
py==1.8.2
pyparsing==2.4.7
PyPika==0.37.6
six==1.14.0
tortoise-orm==0.16.10
typing-extensions==3.7.4.2
wcwidth==0.1.9
yarl==1.4.2
zipp==3.1.0
pytest-aiohttp==0.3.0
nest_asyncio==1.3.3
asynctest==0.13.0
sentry-sdk==0.14.1
pytest-mock==3.1.1
pytz